+++ src/main/java/lotto/exception/budget/BudgetRangeException.java
@@ -1,9 +1,9 @@
-package lotto.exception.money;
+package lotto.exception.budget;
 
-public class MoneyRangeException extends IllegalArgumentException{
+public class BudgetRangeException extends IllegalArgumentException{
     private static final String ERROR_MSG = "[ERROR] 금액은 최소 1000원 이상이어야 합니다.";
 
-    public MoneyRangeException() {
+    public BudgetRangeException() {
         super(ERROR_MSG);
     }
 }
