+++ src/test/java/lotto/model/BudgetTest.java
@@ -1,11 +1,10 @@
-package lotto;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.exception.money.MoneyDivisionException;
-import lotto.exception.money.MoneyRangeException;
-import lotto.model.Budget;
+import lotto.exception.budget.BudgetDivisionException;
+import lotto.exception.budget.BudgetRangeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,14 +13,14 @@ public class BudgetTest {
     @Test
     void createMoneyOutOfRange(){
         assertThatThrownBy(() -> new Budget(900))
-                .isInstanceOf(MoneyRangeException.class);
+                .isInstanceOf(BudgetRangeException.class);
     }
 
     @DisplayName("금액이 1000의 배수가 아니면 예외가 발생한다.")
     @Test
     void createMoneyNotDividedBy1000(){
         assertThatThrownBy(() -> new Budget(1100))
-                .isInstanceOf(MoneyDivisionException.class);
+                .isInstanceOf(BudgetDivisionException.class);
     }
 
     @DisplayName("로또 수량 개수 확인 테스트")
